# Fix bug #5476: Ltac has an inconsistent view of hypotheses.