Fix for [bug 5476](https://coq.inria.fr/bugs/show_bug.cgi?id=5476). Before merging this, I'd like to run a bench to see whether it has a performance impact.