@ppedrot 2017-04-19
Timings are unchanged.
```
┌────────────────────────┬─────────────────────────┬─────────────────────────────────────┬─────────────────────────────────────┐
│                        │      user time [s]      │             CPU cycles              │          CPU instructions           │
│                        │                         │                                     │                                     │
│           package_name │     NEW     OLD PDIFF   │           NEW           OLD PDIFF   │           NEW           OLD PDIFF   │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-odd_order │ 1402.22 1410.15 -0.56 % │ 3907133076155 3922205685454 -0.38 % │ 6825845352130 6825777404342 +0.00 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│              coq-flocq │   55.86   56.17 -0.55 % │  152875286588  153664195277 -0.51 % │  195924191362  195972329430 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│                 coq-sf │   43.94   44.18 -0.54 % │  119729986048  120116834345 -0.32 % │  155792698037  155860024020 -0.04 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│       coq-fiat-parsers │  529.22  531.69 -0.46 % │ 1443637162945 1450068132993 -0.44 % │ 2056545586490 2056642514353 -0.00 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│        coq-fiat-crypto │ 1760.69 1767.04 -0.36 % │ 4888491717191 4906421150637 -0.37 % │ 8118824139231 8111311056628 +0.09 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│               coq-hott │  234.37  235.20 -0.35 % │  633239809782  634197509203 -0.15 % │ 1075167002749 1075340658407 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│  coq-mathcomp-fingroup │   57.24   57.41 -0.30 % │  158175333778  158488609501 -0.20 % │  222481716234  222602553731 -0.05 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│              coq-color │  620.54  622.28 -0.28 % │ 1717141034684 1726413665372 -0.54 % │ 2160468736953 2160586441614 -0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│         coq-coquelicot │   71.16   71.35 -0.27 % │  195346926494  195636126053 -0.15 % │  245959229716  245585709678 +0.15 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│            coq-unimath │ 1252.86 1255.88 -0.24 % │ 3472335952161 3482725801078 -0.30 % │ 5831240656538 5832679667438 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│               coq-iris │  435.40  436.31 -0.21 % │ 1209664759421 1211975416799 -0.19 % │ 1733290407139 1733068014170 +0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│           coq-compcert │  807.58  808.70 -0.14 % │ 2240294109074 2245632888033 -0.24 % │ 3372805274392 3372951348402 -0.00 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│                coq-vst │ 1503.75 1505.14 -0.09 % │ 4174244652929 4180347810459 -0.15 % │ 6268932392129 6269211694164 -0.00 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│   coq-mathcomp-algebra │  167.96  168.11 -0.09 % │  466260780252  466536315286 -0.06 % │  663882967647  664131096798 -0.04 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│     coq-mathcomp-field │  453.76  453.94 -0.04 % │ 1262219832107 1261471634482 +0.06 % │ 2096429355803 2096713021711 -0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-character │  267.50  267.48 +0.01 % │  743230871098  744385300426 -0.16 % │ 1093133811212 1093308414526 -0.02 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│  coq-mathcomp-solvable │  190.29  189.73 +0.30 % │  527355925200  527381208745 -0.00 % │  766878383090  766969896043 -0.01 % │
├────────────────────────┼─────────────────────────┼─────────────────────────────────────┼─────────────────────────────────────┤
│ coq-mathcomp-ssreflect │   42.80   42.56 +0.56 % │  116125271539  115769920020 +0.31 % │  149366412396  149429297449 -0.04 % │
└────────────────────────┴─────────────────────────┴─────────────────────────────────────┴─────────────────────────────────────┘
NEW = 24e19ef548e827173a67c99ba653718d49cdfa6e
OLD = c86c6558fcf7f8dc4a17aceed24f68f756f28ea9
```